requirements.txt
# Core dependencies
ultralytics>=8.3.0    # YOLO11n model & inference
opencv-python>=4.8.0  # Image I/O, undistortion, display
numpy>=1.24.0         # Array math, coordinate transforms
scipy>=1.11.0         # Signal processing (optional, for future filters)
pyyaml>=6.0           # Configuration file support

# SIL simulation (MuJoCo) - Optional, only needed for simulation testing
# mujoco>=3.0.0         # Physics simulation for closed-loop testing

# Hardware integration (optional)
pyserial>=3.5         # Serial port communication for real gimbal/IMU
flask>=2.3.0          # HTTP API for config hot-reload (optional)
paho-mqtt>=1.6.0      # MQTT telemetry (optional)

# Testing
pytest>=7.4.0         # Test framework
pytest-cov>=4.1.0     # Coverage reporting
pytest-xdist>=3.3.0   # Parallel test execution
pytest-benchmark>=4.0.0  # Performance benchmarking

# Code quality
ruff>=0.1.0           # Linter and formatter
mypy>=1.5.0           # Type checking
safety>=2.3.0         # Security vulnerability scanning
